+++ src/config/mod.rs
@@ -1,9 +1,11 @@
 use crate::cli::Arguments;
 use miette::Diagnostic;
+use nix::poll::{PollFd, PollFlags, PollTimeout, poll};
 use owo_colors::OwoColorize;
 use std::cmp::PartialEq;
 use std::fs::File;
-use std::io::{self, IsTerminal, Read, stdin};
+use std::io::{self, Read, stdin};
+use std::os::fd::AsFd;
 use std::path::{Path, PathBuf};
 use std::{env, fs};
 use thiserror::Error;
@@ -48,6 +50,10 @@ pub enum ConfigError {
     #[error("Cannot read from both file and {}", "--listen-command".magenta())]
     CannotReadBothFileAndListenCommand,
 
+    #[error("Detected input from both {} and from {}", "stdin".magenta(), "file".yellow())]
+    #[diagnostic(severity(Warning))]
+    AmbiguousInput,
+
     #[error("Could not determine input type")]
     CouldNotDetermineInputType,
 
@@ -73,9 +79,9 @@ pub fn get_io_config(args: &Arguments) -> Result<InputOutputConfig, ConfigError>
 }
 
 fn get_input(args: &Arguments) -> Result<Source, ConfigError> {
-    let has_data_from_stdin = !stdin().is_terminal();
+    let std_in_has_data = stdin_has_data();
 
-    if !has_data_from_stdin && args.file_path.is_none() && args.listen_command.is_none() {
+    if !std_in_has_data && args.file_path.is_none() && args.listen_command.is_none() {
         return Err(ConfigError::MissingFilename);
     }
 
@@ -83,7 +89,11 @@ fn get_input(args: &Arguments) -> Result<Source, ConfigError> {
         return Err(ConfigError::CannotReadBothFileAndListenCommand);
     }
 
-    if has_data_from_stdin {
+    if std_in_has_data && args.file_path.is_some() {
+        return Err(ConfigError::AmbiguousInput);
+    }
+
+    if std_in_has_data {
         return Ok(Source::Stdin);
     }
 
@@ -96,6 +106,19 @@ fn get_input(args: &Arguments) -> Result<Source, ConfigError> {
     }
 
     Err(ConfigError::CouldNotDetermineInputType)
+}
+
+fn stdin_has_data() -> bool {
+    let stdin = stdin();
+    let fd = stdin.as_fd();
+    let mut fds = [PollFd::new(fd, PollFlags::POLLIN)];
+    match poll(&mut fds, PollTimeout::ZERO) {
+        Ok(n) if n > 0 => fds[0]
+            .revents()
+            .unwrap_or(PollFlags::empty())
+            .contains(PollFlags::POLLIN),
+        _ => false,
+    }
 }
 
 fn get_output(args: &Arguments, input: &Source) -> OutputTarget {